requirements.txt
openai
openai-whisper
coqui-tts
sounddevice
numpy
scipy
python-dotenv
tk
